separation of responsibilities
