+++ projects/rocprofiler-compute/src/lib/roctx_recordfn/tests/test_roctx_recordfn.cpp
@@ -21,6 +21,8 @@
 #include <thread>
 #include <vector>
 
+using namespace roctx_recordfn::detail;
+
 namespace
 {
